Controller and arm control

+++ src/servo-control.cpp
@@ -1,13 +1,13 @@
 #include "robot.h"
 
 
-void Robot::SetServoAngle1(const int angle)
+void Robot::SetServoAngle1(const float angle)
 {
     servoMainRot = angle;
     servo1.write(angle + servoMainOffset);
 }
 
-void Robot::SetServoAngle2(const int angle) {
+void Robot::SetServoAngle2(const float angle) {
     servoSecondRot = angle;
     servo2.write(angle + servoSecondOffset);
 }
